Add feature variant for voice
