+++ core/build.gradle.kts
@@ -1,16 +1,26 @@
 import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
 
 plugins {
+    java
     `kord-module`
     `kord-sampled-module`
     `kord-publishing`
+}
+
+val voice by sourceSets.creating
+
+configurations {
+    getByName("voiceImplementation") {
+        extendsFrom(implementation.get())
+    }
 }
 
 dependencies {
     api(projects.common)
     api(projects.rest)
     api(projects.gateway)
-    api(projects.voice)
+    "voiceApi"(projects.core)
+    "voiceApi"(projects.voice)
 
     implementation(libs.bundles.common)
 
@@ -23,8 +33,19 @@ dependencies {
     testRuntimeOnly(libs.bundles.test.runtime)
 }
 
-tasks.withType<KotlinCompile> {
-    kotlinOptions {
-        freeCompilerArgs = freeCompilerArgs + CompilerArguments.stdLib
+tasks {
+    withType<KotlinCompile> {
+        kotlinOptions {
+            freeCompilerArgs = freeCompilerArgs + CompilerArguments.stdLib
+        }
+    }
+}
+
+java {
+    registerFeature("voice") {
+        usingSourceSet(voice)
+        withJavadocJar()
+        withSourcesJar()
+        capability("dev.kord", "core-voice", version as String)
     }
 }
