+++ core/src/main/kotlin/behavior/channel/BaseVoiceChannelBehavior.kt
@@ -1,16 +1,11 @@
 package dev.kord.core.behavior.channel
 
 import dev.kord.cache.api.query
-import dev.kord.common.annotation.KordVoice
 import dev.kord.common.exception.RequestException
 import dev.kord.core.Kord
 import dev.kord.core.cache.data.VoiceStateData
 import dev.kord.core.cache.idEq
 import dev.kord.core.entity.VoiceState
-import dev.kord.core.entity.channel.VoiceChannel
-import dev.kord.core.exception.GatewayNotFoundException
-import dev.kord.voice.VoiceConnection
-import dev.kord.voice.VoiceConnectionBuilder
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 
@@ -28,27 +23,4 @@ public interface BaseVoiceChannelBehavior : CategorizableChannelBehavior {
         get() = kord.cache.query<VoiceStateData> { idEq(VoiceStateData::channelId, id) }
             .asFlow()
             .map { VoiceState(it, kord) }
-
-    /**
-     * Connect to this [VoiceChannel] and create a [VoiceConnection] for this voice session.
-     *
-     * @param builder a builder for the [VoiceConnection].
-     * @throws GatewayNotFoundException when there is no associated [dev.kord.gateway.Gateway] for the [dev.kord.core.entity.Guild] this channel is in.
-     * @throws dev.kord.voice.exception.VoiceConnectionInitializationException when there was a problem retrieving voice information from Discord.
-     * @return a [VoiceConnection] representing the connection to this [VoiceConnection].
-     */
-    @KordVoice
-    public suspend fun connect(builder: VoiceConnectionBuilder.() -> Unit): VoiceConnection {
-        val voiceConnection = VoiceConnection(
-            guild.gateway ?: GatewayNotFoundException.voiceConnectionGatewayNotFound(guildId),
-            kord.selfId,
-            id,
-            guildId,
-            builder
-        )
-
-        voiceConnection.connect()
-
-        return voiceConnection
-    }
 }
